+++ core/server/api_container/server/startosis_engine/kurtosis_types/port_spec_test.go
@@ -1,26 +1,28 @@
 package kurtosis_types
 
 import (
+	"fmt"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	"github.com/stretchr/testify/require"
 	"go.starlark.net/starlark"
 	"testing"
 )
 
 const (
-	testProtocol = starlark.String("TCP")
+	testProtocol = kurtosis_core_rpc_api_bindings.Port_TCP
 )
 
-var testPortNumber = starlark.MakeUint(123)
+var testPortNumber = uint32(123)
 
 func TestPortSpec_StringRepresentation(t *testing.T) {
 	portSpec := NewPortSpec(testPortNumber, testProtocol)
-	expectedStr := "port_spec(number=123, protocol=\"TCP\")"
+	expectedStr := `PortSpec(number=123, protocol="TCP")`
 	require.Equal(t, expectedStr, portSpec.String())
 }
 
 func TestPortSpec_PortType(t *testing.T) {
 	portSpec := NewPortSpec(testPortNumber, testProtocol)
-	require.Equal(t, portSpecTypeName, portSpec.Type())
+	require.Equal(t, PortSpecTypeName, portSpec.Type())
 }
 
 func TestPort_Freeze(t *testing.T) {
@@ -35,14 +37,6 @@ func TestPort_TruthValidService(t *testing.T) {
 	require.Equal(t, starlark.Bool(true), portSpec.Truth())
 }
 
-func TestPort_TruthFalsyService(t *testing.T) {
-	portSpec := PortSpec{
-		number:   starlark.Int{},
-		protocol: "",
-	}
-	require.Equal(t, starlark.Bool(false), portSpec.Truth())
-}
-
 func TestPort_HashThrowsError(t *testing.T) {
 	portSpec := NewPortSpec(testPortNumber, testProtocol)
 	hash, err := portSpec.Hash()
@@ -54,7 +48,7 @@ func TestPort_TestValidAttr(t *testing.T) {
 	portSpec := NewPortSpec(testPortNumber, testProtocol)
 	attrValue, err := portSpec.Attr(portNumberAttr)
 	require.Nil(t, err)
-	require.Equal(t, testPortNumber, attrValue)
+	require.Equal(t, starlark.MakeInt(int(testPortNumber)), attrValue)
 }
 
 func TestPort_TestInvalidAttr(t *testing.T) {
@@ -68,4 +62,60 @@ func TestPort_TestAttrNames(t *testing.T) {
 	portSpec := NewPortSpec(testPortNumber, testProtocol)
 	attrNames := portSpec.AttrNames()
 	require.Equal(t, []string{portNumberAttr, portProtocolAttr}, attrNames)
+}
+
+func TestParsePortProtocol_TCP(t *testing.T) {
+	input := "TCP"
+	output, err := parsePortProtocol(input)
+	require.Nil(t, err)
+	require.Equal(t, kurtosis_core_rpc_api_bindings.Port_TCP, output)
+}
+
+func TestParsePortProtocol_UDP(t *testing.T) {
+	input := "UDP"
+	output, err := parsePortProtocol(input)
+	require.Nil(t, err)
+	require.Equal(t, kurtosis_core_rpc_api_bindings.Port_UDP, output)
+}
+
+func TestParsePortProtocol_SCTP(t *testing.T) {
+	input := "SCTP"
+	output, err := parsePortProtocol(input)
+	require.Nil(t, err)
+	require.Equal(t, kurtosis_core_rpc_api_bindings.Port_SCTP, output)
+}
+
+func TestParsePortProtocol_DefaultValue(t *testing.T) {
+	input := ""
+	output, err := parsePortProtocol(input)
+	require.Nil(t, err)
+	require.Equal(t, kurtosis_core_rpc_api_bindings.Port_TCP, output)
+}
+
+func TestParsePortProtocol_Unknown(t *testing.T) {
+	input := "BLAH"
+	output, err := parsePortProtocol(input)
+	require.NotNil(t, err)
+	require.Equal(t, "Port protocol should be one of TCP, SCTP, UDP", err.Error())
+	require.Equal(t, kurtosis_core_rpc_api_bindings.Port_Protocol(-1), output)
+}
+
+func TestValidatePort_Success(t *testing.T) {
+	number := 1234
+	_, err := parsePortNumber(number)
+	require.Nil(t, err)
+}
+
+func TestParsePort_FailurePortNumberInvalid(t *testing.T) {
+	number := 123456
+	_, err := parsePortNumber(number)
+	require.NotNil(t, err)
+	require.Equal(t, fmt.Sprintf("Port number should be in range [%d - %d]", 1, 65535), err.Error())
+}
+
+func TestParsePort_FailurePortNumberInvalidNegativeNumber(t *testing.T) {
+	number := -1
+	_, err := parsePortNumber(number)
+	require.NotNil(t, err)
+	require.Equal(t, fmt.Sprintf("Port number should be in range [%d - %d]", 1, 65535), err.Error())
 }