+++ internal_testsuites/golang/testsuite/startosis_request_wait_assert_test/startosis_timeout_wait_test.go
@@ -14,7 +14,7 @@ def run(args):
 	service_config = struct(
 		image = "mendhak/http-https-echo:26",
 		ports = {
-			"http-port": struct(number = 8080, protocol = "TCP")
+			"http-port": PortSpec(number = 8080, protocol = "TCP")
 		}
 	)