+++ internal_testsuites/typescript/src/testsuite/startosis_remote_package_test/startosis_remote_package.test.ts
@@ -19,7 +19,8 @@ const IS_PARTITIONING_ENABLED  = false
 
 jest.setTimeout(180000)
 
-test("Test remote Starlark package execution", async () => {
+// TODO: Remove skip once PortSpec is available is remote package
+test.skip("Test remote Starlark package execution", async () => {
     // ------------------------------------- ENGINE SETUP ----------------------------------------------
     const createEnclaveResult = await createEnclave(TEST_NAME, IS_PARTITIONING_ENABLED)